Leverage forward declarations to improve compile times (#990)

+++ tesseract_collision/test/benchmarks/fcl_discrete_bvh_benchmarks.cpp
@@ -5,6 +5,7 @@
 #include <tesseract_collision/test_suite/benchmarks/large_dataset_benchmarks.hpp>
 #include <tesseract_collision/test_suite/benchmarks/benchmark_utils.hpp>
 #include <tesseract_collision/fcl/fcl_discrete_managers.h>
+#include <tesseract_geometry/geometry.h>
 
 using namespace tesseract_collision;
 using namespace test_suite;
@@ -74,7 +75,8 @@ int main(int argc, char** argv)
           auto tf = Eigen::Isometry3d::Identity();
           std::string name = "BM_CONTACT_TEST_0_" + checker->getName() + "_" +
                              ContactTestTypeStrings[static_cast<std::size_t>(test_type)] + "_" +
-                             GeometryTypeStrings[type1] + "_" + GeometryTypeStrings[type2];
+                             GeometryTypeStrings[static_cast<std::size_t>(type1)] + "_" +
+                             GeometryTypeStrings[static_cast<std::size_t>(type2)];
           benchmark::RegisterBenchmark(name.c_str(),
                                        BM_CONTACT_TEST_FUNC,
                                        DiscreteBenchmarkInfo(checker,
@@ -101,7 +103,8 @@ int main(int argc, char** argv)
           auto tf = Eigen::Isometry3d::Identity();
           std::string name = "BM_CONTACT_TEST_1_" + checker->getName() + "_" +
                              ContactTestTypeStrings[static_cast<std::size_t>(test_type)] + "_" +
-                             GeometryTypeStrings[type1] + "_" + GeometryTypeStrings[type2];
+                             GeometryTypeStrings[static_cast<std::size_t>(type1)] + "_" +
+                             GeometryTypeStrings[static_cast<std::size_t>(type2)];
           benchmark::RegisterBenchmark(name.c_str(),
                                        BM_CONTACT_TEST_FUNC,
                                        DiscreteBenchmarkInfo(checker,
@@ -129,7 +132,8 @@ int main(int argc, char** argv)
           auto tf = Eigen::Isometry3d::Identity();
           std::string name = "BM_CONTACT_TEST_2_" + checker->getName() + "_" +
                              ContactTestTypeStrings[static_cast<std::size_t>(test_type)] + "_" +
-                             GeometryTypeStrings[type1] + "_" + GeometryTypeStrings[type2];
+                             GeometryTypeStrings[static_cast<std::size_t>(type1)] + "_" +
+                             GeometryTypeStrings[static_cast<std::size_t>(type2)];
           benchmark::RegisterBenchmark(name.c_str(),
                                        BM_CONTACT_TEST_FUNC,
                                        DiscreteBenchmarkInfo(checker,
